+++ src/main/java/com/flab/dduikka/config/FeignClientConfig.java
@@ -3,7 +3,7 @@ package com.flab.dduikka.config;
 import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.Configuration;
 
-import com.flab.dduikka.domain.weather.application.AccuWeatherFeignClient;
+import com.flab.dduikka.domain.weather.client.AccuWeatherFeignClient;
 
 @Configuration
 @EnableFeignClients(basePackageClasses = AccuWeatherFeignClient.class)
